Okay, I've gotten a moderate speedup by optimizing the hell out of ViewAsIconCell. I now cache data, don't make any extraneous copies, and trimmed what ViewAsIconCell did.

+++ NSAttributedString+Truncation.h
@@ -30,5 +30,5 @@
 
 
 @interface NSAttributedString (Truncation)
--(NSAttributedString*)truncateForWidth:(int)inWidth;
+-(NSAttributedString*)truncateForWidth:(float)inWidth;
 @end
